requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cssselect2==0.4.1
da-vinci==0.3.0
defusedxml==0.7.1
Django==3.2.9
django-appconf==1.0.5
django-thumbnails==0.6.0
djangorestframework==3.12.4
easy-thumbnails==2.8
easy-thumbnails-rest==1.1.4
httpie==2.6.0
idna==3.3
lxml==4.6.4
pilkit==2.0
Pillow==8.4.0
Pygments==2.10.0
PySocks==1.7.1
python-magic==0.4.24
pytz==2021.3
reportlab==3.6.2
requests==2.26.0
requests-toolbelt==0.9.1
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
tinycss2==1.1.0
urllib3==1.26.7
webencodings==0.5.1
